unit tests for checkoutWithDiscounts
*float > BigDecimal
* test to make sure the math still maths the same way as the v1 logic
* make sure final price isn't negative & math around that feature
* make sure discount math works out appropriately

+++ src/main/java/com/heb/receiptapi/model/Coupon.java
@@ -1,5 +1,7 @@
 package com.heb.receiptapi.model;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Data;
 import lombok.Value;
@@ -15,5 +17,5 @@ import lombok.extern.jackson.Jacksonized;
 public class Coupon {
     private String couponName;
     private long appliedSku;
-    private float discountPrice;
+    private BigDecimal discountPrice;
 }
