+++ src/test/java/com/heb/receiptapi/repository/CouponRepoImplTest.java
@@ -2,6 +2,8 @@ package com.heb.receiptapi.repository;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.io.IOException;
+import java.math.BigDecimal;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.test.util.ReflectionTestUtils;
 
@@ -11,32 +13,32 @@ public class CouponRepoImplTest {
     @Test
     void testGetDiscountMissingFile() throws IOException {
         ReflectionTestUtils.setField(couponRepoImpl, "couponFilePath", "missing");
-        float actual = couponRepoImpl.getDiscount(0);
-        float expected = 0;
+        BigDecimal actual = couponRepoImpl.getDiscount(0);
+        BigDecimal expected = BigDecimal.valueOf(0);
         assertEquals(expected, actual);
     }
     
     @Test
     void testGetDiscountEmptyFile() throws IOException {
         ReflectionTestUtils.setField(couponRepoImpl, "couponFilePath", "src/test/resources/static/emptycoupons.json");
-        float actual = couponRepoImpl.getDiscount(123456);
-        float expected = 0;
+        BigDecimal actual = couponRepoImpl.getDiscount(123456);
+        BigDecimal expected = BigDecimal.valueOf(0);
         assertEquals(expected, actual);
     }
 
     @Test
     void testGetDiscountNotFound() throws IOException {
         ReflectionTestUtils.setField(couponRepoImpl, "couponFilePath", "src/test/resources/static/coupons.json");
-        float actual = couponRepoImpl.getDiscount(0);
-        float expected = 0;
+        BigDecimal actual = couponRepoImpl.getDiscount(0);
+        BigDecimal expected = BigDecimal.valueOf(0);
         assertEquals(expected, actual);
     }
 
     @Test
     void testGetDiscount() throws IOException {
         ReflectionTestUtils.setField(couponRepoImpl, "couponFilePath", "src/test/resources/static/coupons.json");
-        float actual = couponRepoImpl.getDiscount(123456);
-        float expected = 0.99f;
+        BigDecimal actual = couponRepoImpl.getDiscount(123456);
+        BigDecimal expected = BigDecimal.valueOf(0.99);
         assertEquals(expected, actual);
     }
 }
